move global file names to constants
