+++ src/global-constants.ts
@@ -1,0 +1,8 @@
+// NOTE: These are here temporarily until we find a better home for them
+export const GlobalFileNames = {
+	apiConversationHistory: "api_conversation_history.json",
+	uiMessages: "ui_messages.json",
+	openRouterModels: "openrouter_models.json",
+	mcpSettings: "cline_mcp_settings.json",
+	clineRules: ".clinerules",
+}
